hwmon: ina230: add type casting for negative current
When precision_multiplier has positive value and current has negative
value, the calculation makes wrong result by implicit type casting.

Bug 1311630

Reviewed-on: http://git-master/r/245046
(cherry picked from commit a62ae4ba0b076ee2349ff3503656e8d9812c0e72)
Reviewed-on: http://git-master/r/248236

+++ drivers/hwmon/ina230.c
@@ -460,7 +460,7 @@ static s32 show_current(struct device *dev,
 	current_mA =
 		(current_mA * data->pdata->power_lsb) / data->pdata->divisor;
 	if (data->pdata->precision_multiplier)
-		current_mA /= data->pdata->precision_multiplier;
+		current_mA /= (s16) data->pdata->precision_multiplier;
 
 	return sprintf(buf, "%d mA\n", current_mA);
 }
